pass image to component
