+++ src/components/component-image/ComponentImage.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { ReactNode } from 'react';
 import { SComponentImage } from './styles/SComponentImage';
 import { SInnerRectangle } from './styles/SInnerRectangle';
 
@@ -6,12 +6,15 @@ export type IComponentImageTypes = "open" | "grid" | "list item";
 
 interface IComponentImage {
   type?: IComponentImageTypes;
+  image?: ReactNode;
 }
 
-export const ComponentImage: React.FC<IComponentImage> = ({ type }) => {
+export const ComponentImage: React.FC<IComponentImage> = ({ type, image }) => {
   return (
     <SComponentImage type={type}>
-      <SInnerRectangle type={type} />
+      <SInnerRectangle type={type}>
+        {image}
+      </SInnerRectangle>
     </SComponentImage>
   )
 }
